# packages/runtime-contract/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/campaigns": {
      "post": {
        "operationId": "CampaignController_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Idempotent replay",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Create a text or image campaign draft",
        "tags": [
          "campaigns"
        ]
      },
      "get": {
        "description": "Search and filters are applied before pagination. Results use updatedAt DESC and campaign ID ASC ordering; meta.total counts the filtered dataset.",
        "operationId": "CampaignController_list",
        "parameters": [
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "sessionId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "query",
            "required": false,
            "in": "query",
            "description": "Trimmed case-insensitive literal substring search on campaign name, or exact campaign UUID.",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Comma-separated campaign statuses. Values are ORed and combined with other filters using AND.",
            "style": "form",
            "explode": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "DRAFT",
                  "ACTIVE",
                  "PAUSED",
                  "ARCHIVED"
                ]
              }
            }
          },
          {
            "name": "scheduleType",
            "required": false,
            "in": "query",
            "description": "Comma-separated schedule types. Values are ORed and combined with other filters using AND.",
            "style": "form",
            "explode": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "enum": [
                  "IMMEDIATE",
                  "ONCE"
                ]
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignListDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Search and filter campaigns for allowlisted sessions",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}": {
      "get": {
        "operationId": "CampaignController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Read a campaign",
        "tags": [
          "campaigns"
        ]
      },
      "patch": {
        "operationId": "CampaignController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCampaignDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Update an editable campaign draft",
        "tags": [
          "campaigns"
        ]
      },
      "delete": {
        "description": "Creates a durable tombstone without deleting immutable run, delivery, or message-job audit data.",
        "operationId": "CampaignController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "expectedRevision",
            "required": true,
            "in": "query",
            "description": "Campaign content revision observed before deletion. A stale value returns HTTP 409.",
            "schema": {
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "name": "expectedTargetsRevision",
            "required": true,
            "in": "query",
            "description": "Campaign target-set revision observed before deletion. A stale value returns HTTP 409.",
            "schema": {
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Delete a quiescent campaign from the active workspace",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}/targets": {
      "get": {
        "operationId": "CampaignController_listTargets",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignTargetListDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "List selected group targets with current capability",
        "tags": [
          "campaigns"
        ]
      },
      "put": {
        "operationId": "CampaignController_replaceTargets",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceCampaignTargetsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignTargetListDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Atomically replace all group targets of a campaign draft",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}/targets/apply-group-list": {
      "post": {
        "operationId": "CampaignController_applyGroupListTargets",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyGroupListTargetsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignTargetListDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Atomically replace DRAFT campaign targets from one saved group-list revision",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}/preflight": {
      "post": {
        "operationId": "CampaignController_preflight",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignPreflightRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignPreflightDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Evaluate a campaign without creating a run",
        "tags": [
          "campaigns"
        ]
      }
    },
    "/api/v1/campaigns/{id}/runs": {
      "post": {
        "operationId": "CampaignController_createRun",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCampaignRunDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRunDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "Create an idempotent durable campaign run",
        "tags": [
          "campaigns"
        ]
      },
      "get": {
        "operationId": "CampaignController_listRuns",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CampaignRunListDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeErrorDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "runtime-key": []
          }
        ],
        "summary": "List runs of a campaign",
        "tags": [
          "campaigns"
        ]
      }
    },
...
  },
  "info": {
    "title": "WA Runtime API",
    "description": "Stable, versioned API contract consumed by WA Studio",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "runtime-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Runtime-Key"
      }
    },
    "schemas": {
...
      "RuntimeErrorDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "CAMPAIGN_NOT_FOUND"
          },
          "message": {
            "type": "string"
          },
          "fieldErrors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
...
      "CreateCampaignDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "text": {
            "type": "string",
            "maxLength": 4096,
            "deprecated": true,
            "description": "Legacy text input. Use content instead; exactly one of text or content is required."
          },
          "content": {
            "description": "Typed campaign content. Exactly one of content or the legacy text field is required.",
            "oneOf": [
              {
                "$ref": "#/components/schemas/TextCampaignContentInputDto"
              },
              {
                "$ref": "#/components/schemas/ImageCampaignContentInputDto"
              }
            ],
            "discriminator": {
              "propertyName": "type",
              "mapping": {
                "TEXT": "#/components/schemas/TextCampaignContentInputDto",
                "IMAGE": "#/components/schemas/ImageCampaignContentInputDto"
              }
            }
          },
          "scheduleType": {
            "type": "string",
            "enum": [
              "IMMEDIATE",
              "ONCE"
            ],
            "default": "IMMEDIATE"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time",
            "description": "Required for ONCE. Ignored and returned as null for IMMEDIATE."
          }
        },
        "required": [
          "sessionId",
          "name"
        ]
      },
...
      "CampaignListDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CampaignDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/CampaignListPageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}